requirements.txt
bzip2==1.0.8
ca-certificates==2023.01.10
click==8.0.4
flask==2.2.2
freetype==2.12.1
giflib==5.2.1
icu==58.2
itsdangerous==2.0.1
jinja2==3.1.2
jpeg==9e
lcms2==2.12
leptonica==1.82.0
lerc==3.0
libarchive==3.6.2
libcxx==14.0.6
libdeflate==1.17
libffi==3.4.4
libiconv==1.16
libpng==1.6.39
libtiff==4.5.0
libwebp-base==1.2.4
libwebp==1.2.4
libxml2==2.10.3
lz4-c==1.9.4
markupsafe==2.1.1
ncurses==6.4
openjpeg==2.4.0
openssl==1.1.1t
packaging==23.0
pillow==9.4.0
pip==23.0.1
pytesseract==0.3.10
python==3.10.11
readline==8.2
setuptools==66.0.0
sqlite==3.41.2
tesseract==5.2.0
tk==8.6.12
tzdata==2023c
werkzeug==2.2.3
wheel==0.38.4
xz==5.4.2
zlib==1.2.13
zstd==1.5.5
